+++ test/browser.js
@@ -34,23 +34,34 @@ describe('IPFS data importing tests on the Browser', function () {
     ], done)
   })
 
-  require('./test-builder')(repo)
-  require('./test-flat-builder')
-  require('./test-balanced-builder')
-  require('./test-trickle-builder')
-  require('./test-fixed-size-chunker')
+  // HAMT
+  require('./hamt')
+  require('./hamt-consumable-buffer')
+  require('./hamt-consumable-hash')
 
-  // relies on data in the repo
-  // require('./test-exporter')(repo)
+  // Chunkers
+  require('./chunker-fixed-size')
 
-  require('./test-consumable-buffer')
-  require('./test-consumable-hash')
-  require('./test-hamt')
-  require('./test-importer')(repo)
-  require('./test-importer-flush')(repo)
-  require('./test-import-export')(repo)
-  require('./test-hash-parity-with-go-ipfs')(repo)
-  require('./test-nested-dir-import-export')(repo)
-  require('./test-dirbuilder-sharding')(repo)
-  require('./test-builder-only-hash')(repo)
+  // Graph Builders
+  require('./builder')(repo)
+  require('./builder-flat')
+  require('./builder-balanced')
+  require('./builder-trickle-dag')
+  require('./builder-only-hash')(repo)
+  require('./builder-dir-sharding')(repo)
+
+  // Importer
+  require('./importer')(repo)
+  require('./importer-flush')(repo)
+
+  // Exporter
+  // TODO: make these tests not require data on the repo
+  // require('./exporter')(repo)
+  // require('./exporter-subtree')(repo)
+
+  // Other
+  require('./import-export')(repo)
+  require('./import-export-nested-dir')(repo)
+  require('./hash-parity-with-go-ipfs')(repo)
+  // require('./with-dag-api')
 })
