+++ test/builder-dir-sharding.js
@@ -17,7 +17,7 @@ const setImmediate = require('async/setImmediate')
 const leftPad = require('left-pad')
 
 module.exports = (repo) => {
-  describe('dirbuilder sharding', function () {
+  describe('builder: directory sharding', function () {
     this.timeout(20 * 1000)
 
     let ipldResolver
