+++ test/hamt.js
