+++ test/import-export.js
@@ -8,7 +8,6 @@ const BlockService = require('ipfs-block-service')
 const IPLDResolver = require('ipld-resolver')
 const pull = require('pull-stream')
 
-const randomByteStream = require('./helpers/finite-pseudorandom-byte-stream')
 const unixFSEngine = require('./../')
 const exporter = unixFSEngine.exporter
 
@@ -17,64 +16,6 @@ const strategies = [
   'balanced',
   'trickle'
 ]
-
-module.exports = (repo) => {
-  let bigFile
-  strategies.forEach((strategy) => {
-    const importerOptions = {
-      strategy: strategy
-    }
-
-    describe('import export using ' + strategy + ' builder strategy', () => {
-      let ipldResolver
-
-      before(() => {
-        const bs = new BlockService(repo)
-        ipldResolver = new IPLDResolver(bs)
-      })
-
-      before((done) => {
-        if (bigFile) {
-          return done()
-        }
-
-        pull(
-          randomByteStream(50000000, 8372),
-          pull.collect((err, buffers) => {
-            if (err) {
-              done(err)
-            } else {
-              bigFile = buffers
-              done()
-            }
-          })
-        )
-      })
-
-      it('import and export', (done) => {
-        const path = strategy + '-big.dat'
-        pull(
-          pull.values([{
-            path: path,
-            content: pull.values(bigFile)
-          }]),
-          unixFSEngine.importer(ipldResolver, importerOptions),
-          pull.map((file) => {
-            expect(file.path).to.be.eql(path)
-
-            return exporter(file.multihash, ipldResolver)
-          }),
-          pull.flatten(),
-          pull.collect((err, files) => {
-            expect(err).to.not.exist()
-            expect(files[0].size).to.be.eql(bigFile.reduce(reduceLength, 0))
-            fileEql(files[0], Buffer.concat(bigFile), done)
-          })
-        )
-      })
-    })
-  })
-}
 
 function fileEql (f1, f2, done) {
   pull(
@@ -86,7 +27,7 @@ function fileEql (f1, f2, done) {
 
       try {
         if (f2) {
-          expect(Buffer.concat(data)).to.be.eql(f2)
+          expect(Buffer.concat(data)).to.eql(f2)
         } else {
           expect(data).to.exist()
         }
@@ -100,4 +41,44 @@ function fileEql (f1, f2, done) {
 
 function reduceLength (acc, chunk) {
   return acc + chunk.length
+}
+
+module.exports = (repo) => {
+  const bigFile = Buffer.alloc(5000000, 'a')
+
+  strategies.forEach((strategy) => {
+    const importerOptions = { strategy: strategy }
+
+    describe('import and export with builder: ' + strategy, () => {
+      let ipldResolver
+
+      before(() => {
+        const bs = new BlockService(repo)
+        ipldResolver = new IPLDResolver(bs)
+      })
+
+      // TODO fix pull-block https://github.com/dignifiedquire/pull-block/pull/10
+      it('import and export', (done) => {
+        const path = strategy + '-big.dat'
+        pull(
+          pull.values([{
+            path: path,
+            content: pull.values(bigFile)
+          }]),
+          unixFSEngine.importer(ipldResolver, importerOptions),
+          pull.map((file) => {
+            expect(file.path).to.eql(path)
+
+            return exporter(file.multihash, ipldResolver)
+          }),
+          pull.flatten(),
+          pull.collect((err, files) => {
+            expect(err).to.not.exist()
+            expect(files[0].size).to.be.eql(bigFile.reduce(reduceLength, 0))
+            fileEql(files[0], Buffer.concat(bigFile), done)
+          })
+        )
+      })
+    })
+  })
 }
