+++ test/importer.js
@@ -160,7 +160,7 @@ module.exports = (repo) => {
 
     const expected = extend({}, defaultResults, strategies[strategy])
 
-    describe(strategy + ' importer', function () {
+    describe('importer: ' + strategy, function () {
       this.timeout(20 * 1000)
 
       let ipldResolver
