+++ test/node.js
@@ -4,13 +4,14 @@
 const ncp = require('ncp').ncp
 const rimraf = require('rimraf')
 const path = require('path')
+const os = require('os')
 const IPFSRepo = require('ipfs-repo')
 const mkdirp = require('mkdirp')
 const series = require('async/series')
 
 describe('IPFS UnixFS Engine', () => {
   const repoExample = path.join(process.cwd(), '/test/test-repo')
-  const repoTests = path.join(process.cwd(), '/test/repo-tests' + Date.now())
+  const repoTests = path.join(os.tmpdir(), '/unixfs-tests-' + Date.now())
 
   const repo = new IPFSRepo(repoTests)
 
@@ -36,22 +37,33 @@ describe('IPFS UnixFS Engine', () => {
     ], done)
   })
 
-  require('./test-builder')(repo)
-  require('./test-flat-builder')
-  require('./test-balanced-builder')
-  require('./test-trickle-builder')
-  require('./test-fixed-size-chunker')
-  require('./test-consumable-buffer')
-  require('./test-consumable-hash')
-  require('./test-hamt')
-  require('./test-exporter')(repo)
-  require('./test-export-subtree')(repo)
-  require('./test-importer')(repo)
-  require('./test-importer-flush')(repo)
-  require('./test-import-export')(repo)
-  require('./test-hash-parity-with-go-ipfs')(repo)
-  require('./test-nested-dir-import-export')(repo)
-  require('./test-dirbuilder-sharding')(repo)
-  require('./test-dag-api')
-  require('./test-builder-only-hash')(repo)
+  // HAMT
+  require('./hamt')
+  require('./hamt-consumable-buffer')
+  require('./hamt-consumable-hash')
+
+  // Chunkers
+  require('./chunker-fixed-size')
+
+  // Graph Builders
+  require('./builder')(repo)
+  require('./builder-flat')
+  require('./builder-balanced')
+  require('./builder-trickle-dag')
+  require('./builder-only-hash')(repo)
+  require('./builder-dir-sharding')(repo)
+
+  // Importer
+  require('./importer')(repo)
+  require('./importer-flush')(repo)
+
+  // Exporter
+  require('./exporter')(repo)
+  require('./exporter-subtree')(repo)
+
+  // Other
+  require('./import-export')(repo)
+  require('./import-export-nested-dir')(repo)
+  require('./hash-parity-with-go-ipfs')(repo)
+  require('./with-dag-api')
 })
