+++ test/test-repo/version
@@ -1,1 +1,1 @@
-5
+6
